+++ schedule/src/main/java/itcast/blog/application/VelogCrawlingService.java
@@ -7,11 +7,11 @@ import itcast.blog.client.VelogHttpClient;
 import itcast.blog.parser.VelogDataParser;
 import itcast.blog.repository.BlogRepository;
 import itcast.domain.blog.Blog;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 @Service
 @Slf4j
@@ -64,12 +64,5 @@ public class VelogCrawlingService {
                     gptService.updateBlogBySummaryContent(request, savedId.getId());
                 }
         );
-    }
-
-    @Scheduled(cron = "${scheduler.velog.crawling}")
-    public void velogCrawling() {
-        log.info("Velog Crawling Start ...");
-        crawlBlogs();
-        log.info("Velog Crawling & Save!");
     }
 }
