+++ schedule/src/main/java/itcast/blog/scheduler/BlogCrawlSchedule.java
@@ -1,0 +1,32 @@
+package itcast.blog.scheduler;
+
+import itcast.blog.application.VelogCrawlingService;
+import itcast.blog.application.YozmCrawlingService;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.stereotype.Component;
+
+@Slf4j
+@Component
+@RequiredArgsConstructor
+public class BlogCrawlSchedule {
+
+    private final VelogCrawlingService velogCrawlingService;
+    private final YozmCrawlingService yozmCrawlingService;
+
+
+    @Scheduled(cron = "${scheduler.velog.crawling}")
+    public void velogCrawling() {
+        log.info("Velog Crawling Start ...");
+        velogCrawlingService.crawlBlogs();
+        log.info("Velog Crawling & Save!");
+    }
+
+    @Scheduled(cron = "${scheduler.yozm.crawling}")
+    public void yozmCrawling() {
+        log.info("Yozm Crawling Start ...");
+        yozmCrawlingService.crawlBlogs();
+        log.info("Yozm Crawling & Save!");
+    }
+}
